Updated a few things
